extended sample wizard
git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@262 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ MGX-Wizards/src/main/java/de/cebitec/mgx/gui/wizard/sample/SampleWizardDescriptor.java
@@ -48,7 +48,7 @@ public class SampleWizardDescriptor extends WizardDescriptor {
         
         sample.setCollectionDate((Date)getProperty(SampleVisualPanel1.PROP_COLLECTIONDATE))
                 .setMaterial((String)getProperty(SampleVisualPanel2.PROP_MATERIAL))
-                .setTemperature((Integer)getProperty(SampleVisualPanel2.PROP_TEMPERATURE))
+                .setTemperature((Double)getProperty(SampleVisualPanel2.PROP_TEMPERATURE))
                 .setVolume((Integer)getProperty(SampleVisualPanel2.PROP_VOLUME))
                 .setVolumeUnit((String)getProperty(SampleVisualPanel2.PROP_VOLUME_UNIT));
         return sample;
